+++ src/components/Plugin/Plugin.jsx
@@ -66,9 +66,13 @@ export class Plugin extends Component {
 
   onStarClicked = () => {
     if (this.isLoggedIn()) {
-      return this.isFavorite() ? this.unfavPlugin() : this.favPlugin();
+      if (this.isFavorite()) 
+        this.unfavPlugin();
+      else 
+        this.favPlugin();
     }
-    return this.showNotifications(new Error('You need to be logged in!'))
+    else
+      this.showNotifications(new Error('You need to be logged in!'))
   }
 
   favPlugin = async () => {
@@ -136,9 +140,13 @@ export class Plugin extends Component {
   async fetchPluginVersions(name) {
     try {
       const versions = await this.client.getPlugins({ limit: 10e6, name });
-      return this.setState((prevState) => 
-        ({ pluginData: { ...prevState.pluginData, versions: versions.data } })
-      );
+      return this.setState((prevState) => ({ 
+        pluginData: { 
+          ...prevState.pluginData, 
+          versions: versions.data, 
+          url: versions.url,
+        } 
+      }));
     } catch (e) {
       this.showNotifications(new HttpApiCallError(e));
       return e
